Reorganize nunit.framework.tests namespaces
