+++ src/NUnitFramework/tests/Constraints/DictionaryContainsKeyConstraintTests.cs
@@ -27,7 +27,7 @@ using System.Collections.Generic;
 using NUnit.Framework.Constraints;
 using NUnit.Framework.Internal;
 
-namespace NUnit.Framework.Tests.Constraints
+namespace NUnit.Framework.Constraints
 {
     [TestFixture]
     public class DictionaryContainsKeyConstraintTests
