+++ src/NUnitFramework/tests/Internal/TestNameGeneratorTests.cs
@@ -25,7 +25,7 @@ using System;
 using System.Reflection;
 using NUnit.Compatibility;
 
-namespace NUnit.Framework.Internal.Tests
+namespace NUnit.Framework.Internal
 {
     public class TestNameGeneratorTests
     {
@@ -45,14 +45,14 @@ namespace NUnit.Framework.Internal.Tests
 
         [TestCase("FIXED", ExpectedResult = "FIXED")]
         [TestCase("{m}", ExpectedResult = "TestMethod")]
-        [TestCase("{n}", ExpectedResult = "NUnit.Framework.Internal.Tests")]
+        [TestCase("{n}", ExpectedResult = "NUnit.Framework.Internal")]
         [TestCase("{c}", ExpectedResult = "TestNameGeneratorTests")]
-        [TestCase("{C}", ExpectedResult = "NUnit.Framework.Internal.Tests.TestNameGeneratorTests")]
-        [TestCase("{M}", ExpectedResult = "NUnit.Framework.Internal.Tests.TestNameGeneratorTests.TestMethod")]
+        [TestCase("{C}", ExpectedResult = "NUnit.Framework.Internal.TestNameGeneratorTests")]
+        [TestCase("{M}", ExpectedResult = "NUnit.Framework.Internal.TestNameGeneratorTests.TestMethod")]
         [TestCase("{m}_SpecialCase", ExpectedResult = "TestMethod_SpecialCase")]
-        [TestCase("{n}.{c}.{m}", ExpectedResult = "NUnit.Framework.Internal.Tests.TestNameGeneratorTests.TestMethod")]
+        [TestCase("{n}.{c}.{m}", ExpectedResult = "NUnit.Framework.Internal.TestNameGeneratorTests.TestMethod")]
         [TestCase("{x}", ExpectedResult = "{x}")]
-        [TestCase("{n}.{c.{m}", ExpectedResult = "NUnit.Framework.Internal.Tests.{c.{m}")]
+        [TestCase("{n}.{c.{m}", ExpectedResult = "NUnit.Framework.Internal.{c.{m}")]
         [TestCase("{m}{a}", ExpectedResult = "TestMethod")]
         [TestCase("{i}", ExpectedResult="THE_ID")]
         public string SimpleTestNames(string pattern)
@@ -87,12 +87,12 @@ namespace NUnit.Framework.Internal.Tests
 
         [TestCase("FIXED", ExpectedResult="FIXED")]
         [TestCase("{m}",   ExpectedResult="GenericTest<T,U,V>")]
-        [TestCase("{n}", ExpectedResult = "NUnit.Framework.Internal.Tests")]
+        [TestCase("{n}", ExpectedResult = "NUnit.Framework.Internal")]
         [TestCase("{c}", ExpectedResult = "TestNameGeneratorTests")]
-        [TestCase("{C}", ExpectedResult = "NUnit.Framework.Internal.Tests.TestNameGeneratorTests")]
-        [TestCase("{M}", ExpectedResult = "NUnit.Framework.Internal.Tests.TestNameGeneratorTests.GenericTest<T,U,V>")]
+        [TestCase("{C}", ExpectedResult = "NUnit.Framework.Internal.TestNameGeneratorTests")]
+        [TestCase("{M}", ExpectedResult = "NUnit.Framework.Internal.TestNameGeneratorTests.GenericTest<T,U,V>")]
         [TestCase("{m}_SpecialCase", ExpectedResult = "GenericTest<T,U,V>_SpecialCase")]
-        [TestCase("{n}.{c}.{m}", ExpectedResult = "NUnit.Framework.Internal.Tests.TestNameGeneratorTests.GenericTest<T,U,V>")]
+        [TestCase("{n}.{c}.{m}", ExpectedResult = "NUnit.Framework.Internal.TestNameGeneratorTests.GenericTest<T,U,V>")]
         public string GenericTestNames(string pattern)
         {
             return new TestNameGenerator(pattern).GetDisplayName(_genericTest);
@@ -102,7 +102,7 @@ namespace NUnit.Framework.Internal.Tests
         [TestCase("{xy}", ExpectedResult = "{xy}")]
         [TestCase("{x:}", ExpectedResult = "{x:}")]
         [TestCase("{x:50}", ExpectedResult = "{x:50}")]
-        [TestCase("{n}.{c.{m}", ExpectedResult = "NUnit.Framework.Internal.Tests.{c.{m}")]
+        [TestCase("{n}.{c.{m}", ExpectedResult = "NUnit.Framework.Internal.{c.{m}")]
         [TestCase("{m}{a:X}", ExpectedResult = "TestMethod{a:X}")]
         [TestCase("{m}{0:X}", ExpectedResult = "TestMethod{0:X}")]
         [TestCase("{m}{a:}", ExpectedResult = "TestMethod{a:}")]
