+++ src/NUnitFramework/tests/TextOutputTests.cs
@@ -27,7 +27,7 @@ using NUnit.Framework.Internal;
 using NUnit.TestData;
 using NUnit.TestUtilities;
 
-namespace NUnit.Framework.Tests
+namespace NUnit.Framework
 {
     public class TextOutputTests : ITestListener
     {
