+++ src/NUnitFramework/tests/nunit.framework.tests-2.0.csproj
@@ -6,7 +6,7 @@
     <ProjectGuid>{EA9871E3-9972-4476-821E-3961C28125E8}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>NUnit.Framework.Tests</RootNamespace>
+    <RootNamespace>NUnit.Framework</RootNamespace>
     <AssemblyName>nunit.framework.tests</AssemblyName>
     <TargetFrameworkVersion>v2.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
@@ -328,7 +328,6 @@
     <Compile Include="Constraints\ExactCountConstraintTests.cs" />
     <Compile Include="TestUtilities\UniqueValues.cs" />
     <Compile Include="TextOutputTests.cs" />
-    <Compile Include="ThrowsTests.cs" />
     <Compile Include="WorkItemTests.cs" />
   </ItemGroup>
   <ItemGroup>
